server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
email-validator==2.1.0
lxml[html_clean]==4.9.3
newspaper3k==0.2.8
readability-lxml==0.8.1
beautifulsoup4==4.12.2
requests==2.31.0
Pillow==10.1.0
moviepy==1.0.3
ffmpeg-python==0.2.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
aiohttp==3.9.1
langdetect==1.0.9